refactor: move files and extract drug from pharmacy.js
