+++ src/pharmacy.js
@@ -1,0 +1,8 @@
+export class Pharmacy {
+  constructor(drugs = []) {
+    this.drugs = drugs;
+  }
+  updateBenefitValue() {
+    return this.drugs.map(drug => drug.updateBenefitValue());
+  }
+}
